% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.rusb}
\alias{predict.rusb}
\title{rusboost}
\usage{
\method{predict}{rusb}(object, newdata, newmfinal = length(object$trees), ...)
}
\arguments{
\item{object}{Fitted model from \code{rusb}.}

\item{newdata}{A new input data frame.}

\item{newmfinal}{Number of trees in boosting object to be used in prediction. By default, all the trees in the object are used.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Returns an object of class \code{predict.boosting}:

\code{formula}  original formula used

\code{votes}    the weighted number of trees which voted for each class

\code{prob}     probability measure calculated using the proportion of votes for each class

\code{class}    predicted class

\code{confusion} confusion matrix

\code{error}    average error
}
\description{
Make predictions using a fitted RUSBoost model
}
\keyword{rusb}

