% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rusb}
\alias{rusb}
\title{rusboost}
\usage{
rusb(formula, data, boot = FALSE, iters = 100, coeflearn = "Breiman",
  control, sampleFraction, idx)
}
\arguments{
\item{formula}{Formula for \code{rpart} models.}

\item{data}{Input dataset.}

\item{boot}{Use bootstrap sampling? (After undersampling phase)}

\item{iters}{How many classifiers to train.}

\item{coeflearn}{From \code{adabag}: "if 'Breiman'(by default), alpha=1/2ln((1-err)/err) is used. If 'Freund' alpha=ln((1-err)/err) is used. In both cases the AdaBoost.M1 algorithm is used and alpha is the weight updating coefficient. On the other hand, if coeflearn is 'Zhu' the SAMME algorithm is implemented with alpha=ln((1-err)/err)+ ln(nclasses-1)."}

\item{control}{Parameters used in \code{rpart} calls.}

\item{sampleFraction}{Percentage of each iteration's sample which should be in the \emph{negative} class. Required.}

\item{idx}{Logical index vector indicating which observations in \code{data} are negative.}
}
\description{
Implement RUSBoost algorithm to deal with unbalanced classes in classification problems
}
\examples{
idx <- data$outcome=="N"
test.rusboost <- rusb(outcome ~ ., data = testdata, boot = F, iters = 20, sampleFraction = .01, idx = idx)
}
\keyword{rusb}

